+++ Kotlin/Demo/app/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "io.chirp.connectdemoapp"
+        applicationId "io.chirp.sdkdemoapp"
         minSdkVersion 15
         targetSdkVersion 28
         versionCode 1
